requirements.txt
python-dotenv
llama-index
llama-index-llms-openai==0.1.31
llama-index-llms-ollama==0.1.3
llama-index-llms-gemini==0.1.7
llama-index-embeddings-ollama==0.1.2
llama-index-tools-duckduckgo==0.1.0
llama-index-vector-stores-chroma==0.1.5
# llama-index-llms-llama-cpp==0.1.4
# vllm==0.4.1
fastapi==0.112.2
easydict==1.9.0
feedparser==6.0.11
schedule==1.2.1
sendgrid==6.11.0
markdown==3.7.0
pyvis==0.3.2
icecream==2.1.3
llama-parse==0.4.9
llama-index-readers-file==0.1.33
llama-index-vector-stores-qdrant==0.2.17
llama-index-vector-stores-elasticsearch
pydantic_settings==2.5.2
llama-index-postprocessor-rankgpt-rerank
llama-index-agent-openai-legacy
elasticsearch==8.15.1
langchain==0.3.0
theflow==0.8.6
PyMuPDF==1.24.10
python-docx==1.1.2
openpyxl==3.1.5
# python-magic-bin==0.4.14
# additional
llama-index-callbacks-aim==0.3.0